+++ features/CommandContext.php
@@ -35,13 +35,15 @@ class CommandContext extends CommonContext
                 break;
             case "cocorico:bookings:expire":
                 $application->add(new ExpireBookingsCommand());
-                $arguments["--delay"] = 0;//cocorico.booking.expiration_delay
+                $arguments["--expiration-delay"] = 0;//cocorico.booking.expiration_delay
+                $arguments["--acceptation-delay"] = 0;//cocorico.booking.acceptation_delay
                 $arguments["--test"] = true;
                 break;
             case "cocorico:bookings:alertExpiring":
                 $application->add(new AlertExpiringBookingsCommand());
-                $arguments["--delay"] = 0;//cocorico.booking.expiration_delay
-                $arguments["--alert_delay"] = 0;//cocorico.booking.alert_expiration_delay
+                $arguments["--alert-delay"] = 0;//cocorico.booking.alert_expiration_delay
+                $arguments["--expiration-delay"] = 0;//cocorico.booking.expiration_delay
+                $arguments["--acceptation-delay"] = 0;//cocorico.booking.acceptation_delay
                 $arguments["--test"] = true;
                 break;
             case "cocorico:bookings:alertImminent":